Migrating from JUnit to TestNG.  Adding the com.wildstartech.wfa.logistics package.  Making some changes to the com.wildstartech.wfa.dao API based on some usage patterns I've observed while writing an implementation for that package running under Google App Engine.

+++ trunk/src/main/java/com/wildstartech/wfa/finance/Charge.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2001 - 2013 Wildstar Technologies, LLC.
+ * Copyright (c) 2001 - 2014 Wildstar Technologies, LLC.
  *
  * This file is part of Wildstar Foundation Architecture.
  *
@@ -35,8 +35,8 @@
  * If you need additional information or have any questions, please contact:
  *
  *      Wildstar Technologies, LLC.
- *      1453 Riverview Run Lane
- *      Suwanee, GA 30024
+ *      63 The Greenway Loop
+ *      Panama City Beach, FL 32413
  *      USA
  *
  *      derek.berube@wildstartech.com
@@ -44,13 +44,15 @@
  */
 package com.wildstartech.wfa.finance;
 
+import java.math.BigDecimal;
+
 public interface Charge {
     public String getDescription();
 
     public void setDescription(String description)
             throws ChargeDescriptionTooLongException;
 
-    public float getAmount();
+    public BigDecimal getAmount();
 
-    public void setAmount(float amount);
+    public void setAmount(BigDecimal amount);
 }
